backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
replicate>=0.25.0
mediapipe>=0.10.9
Pillow>=10.2.0
numpy>=1.26.0
python-dotenv>=1.0.0
httpx>=0.26.0
opencv-python-headless>=4.9.0
